refactor: User route handlers and business logic in controllers
